picoctf/mind_your_ps_and_qs/pyproject.toml
[tool.poetry]
name = "mind_your_ps_and_qs"
version = "0.1.0"
description = ""
authors = ["Antonio Gutierrez <chibby0ne@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
neovim = "^0.3.1"
factordb-python = "^1.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
